style: EditableMemo, TagManager 안에 있던 내부 컴퍼넌트 파일 분리와 index 파일 생성
